+++ src/TrackRowInfoVis.js
@@ -19,8 +19,8 @@ const fieldTypeToVisComponent = {
  * @prop {number} top The top position of this view.
  * @prop {number} width The width of this view.
  * @prop {number} height The height of this view.
- * @prop {object[]} rowInfo Array of JSON objects, one object for each sample, without filtering/sorting based on selected rows.
- * @prop {object[]} transformedRowInfo The `rowInfo` array after transforming by filtering and sorting according to the selected rows.
+ * @prop {object[]} rowInfo The array of JSON Object containing row information.
+ * @prop {object[]} transformedRowInfo The `rowInfo` array after aggregating, filtering, and sorting rows.
  * @prop {object} fieldInfo The name and type of data field.
  * @prop {boolean} isLeft Is this view on the left side of the track?
  * @prop {object} rowSort The options for sorting rows.
@@ -156,8 +156,8 @@ export default function TrackRowInfoVis(props) {
                     isLeft,
                     isShowControlButtons: isHovering,
                     fieldInfo,
-                    rowInfo,
                     transformedRowInfo,
+                    rowInfo,
                     titleSuffix,
                     sortInfo,
                     filterInfo,